Plain CCS of an compiled _reboot.scss
For anyone, who needs to override _reboot.scss, but has no access to that.
Delete the entry _.scss in your mapping files and include this css-file. Prevents overriding iframe background colors, or in general prevents overwriting of any body-attributes, when you include this .css-file instead of the scss (via map). Because mapping can be invoked after you defined your styles. With this file you can exactle place the order of the stylesheets.